refactor: LottoShop에서 생성하던 로또 리스트를 LottoFactory로 이동
