+++ src/main/java/lotto/domain/LottoFactory.java
@@ -17,6 +17,17 @@ public class LottoFactory {
     private LottoFactory() {
     }
 
+    public static Lottos generateLottos(Money money, List<String> inputManualLottoNumbers) {
+        List<Lotto> lottos = new ArrayList<>();
+        for (String s : inputManualLottoNumbers) {
+            lottos.add(LottoFactory.manualGenerator(s));
+        }
+        for (int i=0; i<money.getAutoCount(); i++) {
+            lottos.add(LottoFactory.autoGenerator());
+        }
+        return new Lottos(lottos);
+    }
+
     public static Lotto autoGenerator() {
         Collections.shuffle(LOTTO_NUMBERS);
         return new Lotto(
